Updated script file for easier use, moved walk and dialog event creation to their respective classes - vchadha
