+++ src/com/base/game/levels/LevelManager.java
@@ -28,7 +28,7 @@ public class LevelManager {
         EmptyLevel level1 = new EmptyLevel("res/assets/levelBack.png", true);
 
         Boss boss = new Boss(Display.getWidth() / 2 - 35, Display.getHeight() - 150, 70, 70, "", 3f,60, 8);
-        BossLevel level2 = new BossLevel("res/assets/bossBack.png", boss);
+        BossLevel level2 = new BossLevel("res/assets/bossBack.png", boss, "res/scripts/cutsceneTest1.bsh");
 
         EmptyLevel endGame = new EmptyLevel("res/assets/thankYouForWatching.png", false);
