+++ src/com/base/game/scenes/Event.java
@@ -3,22 +3,15 @@ package com.base.game.scenes;
 import java.util.concurrent.Callable;
 
 public class Event {
-    private String type;
     private String action;
     private boolean isOver;
     private Callable<Boolean> runnable;
 
-    public Event(String type, String action, Callable<Boolean> callback) {
-        this.type = type;
+    public Event(String action, Callable<Boolean> callback) {
         this.action = action;
 
         isOver = false;
         runnable = callback;
-    }
-
-    private int value() {
-        String code = type.toLowerCase() + action.toLowerCase();
-        return code.hashCode();
     }
 
     public void exec() {
@@ -31,20 +24,5 @@ public class Event {
 
     public boolean isOver() {
         return isOver;
-    }
-
-    @Override
-    public int hashCode() {
-        return value();
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj instanceof Event) {
-            Event event = (Event) obj;
-            return (this.type.equals(event.type) && this.action.equals(event.action));
-        }
-
-        return false;
     }
 }
